refactor dpm campaign assignment views
